refactor(epics): extract complexity to pass fallow gate
- server.ts: extract resolveLandTarget helper from handleEpicsCompletedLand
  (26→≤20 cyclomatic, 21→≤15 cognitive)
- IntegratedEpicRow: extract actions into IntegratedEpicLanding child
  (template cognitive 72→≤60)
